+++ napi/tests/resolver.test.mjs
@@ -244,14 +244,14 @@ test('resolve pnpm package', () => {
     styledComponents.path,
     join(
       rootDir,
-      'node_modules/.pnpm/styled-components@6.1.1_react-dom@19.0.0_react@19.0.0__react@19.0.0/node_modules/styled-components/dist/styled-components.browser.cjs.js',
+      'node_modules/.pnpm/styled-components@6.1.16_react-dom@19.0.0_react@19.0.0__react@19.0.0/node_modules/styled-components/dist/styled-components.browser.cjs.js',
     ),
   );
 
   const react = resolver.sync(
     join(
       rootDir,
-      'node_modules/.pnpm/styled-components@6.1.1_react-dom@19.0.0_react@19.0.0__react@19.0.0/node_modules/styled-components',
+      'node_modules/.pnpm/styled-components@6.1.16_react-dom@19.0.0_react@19.0.0__react@19.0.0/node_modules/styled-components',
     ),
     'react',
   );