+++ src/error.rs
@@ -109,12 +109,17 @@ impl ResolveError {
 
     #[must_use]
     #[cfg(feature = "fs_cache")]
-    pub fn from_serde_json_error(path: PathBuf, error: &serde_json::Error) -> Self {
+    pub fn from_serde_json_error(
+        path: PathBuf,
+        error: &serde_json::Error,
+        content: Option<String>,
+    ) -> Self {
         Self::JSON(JSONError {
             path,
             message: error.to_string(),
             line: error.line(),
             column: error.column(),
+            content,
         })
     }
 }
@@ -133,6 +138,7 @@ pub struct JSONError {
     pub message: String,
     pub line: usize,
     pub column: usize,
+    pub content: Option<String>,
 }
 
 #[derive(Debug, Clone, Error)]