+++ src/fs_cache.rs
@@ -127,7 +127,13 @@ impl<Fs: FileSystem> Cache for FsCache<Fs> {
                 };
                 PackageJsonSerde::parse(package_json_path.clone(), real_path, &package_json_string)
                     .map(|package_json| Some((path.clone(), (Arc::new(package_json)))))
-                    .map_err(|error| ResolveError::from_serde_json_error(package_json_path, &error))
+                    .map_err(|error| {
+                        ResolveError::from_serde_json_error(
+                            package_json_path,
+                            &error,
+                            Some(package_json_string),
+                        )
+                    })
             })
             .cloned();
         // https://github.com/webpack/enhanced-resolve/blob/58464fc7cb56673c9aa849e68e6300239601e615/lib/DescriptionFileUtils.js#L68-L82
@@ -176,7 +182,11 @@ impl<Fs: FileSystem> Cache for FsCache<Fs> {
             .map_err(|_| ResolveError::TsconfigNotFound(path.to_path_buf()))?;
         let mut tsconfig = TsConfigSerde::parse(root, &tsconfig_path, &mut tsconfig_string)
             .map_err(|error| {
-                ResolveError::from_serde_json_error(tsconfig_path.to_path_buf(), &error)
+                ResolveError::from_serde_json_error(
+                    tsconfig_path.to_path_buf(),
+                    &error,
+                    Some(tsconfig_string),
+                )
             })?;
         callback(&mut tsconfig)?;
         tsconfig.expand_template_variables();