+++ src/options.rs
@@ -46,6 +46,12 @@ pub struct ResolveOptions {
     /// Default `["package.json"]`
     pub description_files: Vec<String>,
 
+    /// Whether the resolver should check for the presence of a .pnp.cjs file up the dependency tree.
+    ///
+    /// Default `true`
+    #[cfg(feature = "yarn_pnp")]
+    pub enable_pnp: bool,
+
     /// Set to [EnforceExtension::Enabled] for [ESM Mandatory file extensions](https://nodejs.org/api/esm.html#mandatory-file-extensions).
     ///
     /// If `enforce_extension` is set to [EnforceExtension::Enabled], resolution will not allow extension-less files.
@@ -118,12 +124,6 @@ pub struct ResolveOptions {
     /// Default `["node_modules"]`
     pub modules: Vec<String>,
 
-    /// A manifest loaded from pnp::load_pnp_manifest.
-    ///
-    /// Default `None`
-    #[cfg(feature = "yarn_pnp")]
-    pub pnp_manifest: Option<pnp::Manifest>,
-
     /// Resolve to a context instead of a file.
     ///
     /// Default `false`
@@ -168,7 +168,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::ResolveOptions;
+    /// use unrspack_resolver::ResolveOptions;
     ///
     /// let options = ResolveOptions::default().with_condition_names(&["bar"]);
     /// assert_eq!(options.condition_names, vec!["bar".to_string()])
@@ -182,7 +182,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::ResolveOptions;
+    /// use unrspack_resolver::ResolveOptions;
     ///
     /// let options = ResolveOptions::default().with_builtin_modules(false);
     /// assert_eq!(options.builtin_modules, false)
@@ -198,7 +198,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::ResolveOptions;
+    /// use unrspack_resolver::ResolveOptions;
     /// use std::path::{Path, PathBuf};
     ///
     /// let options = ResolveOptions::default().with_root("foo");
@@ -215,7 +215,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::ResolveOptions;
+    /// use unrspack_resolver::ResolveOptions;
     /// use std::path::{Path, PathBuf};
     ///
     /// let options = ResolveOptions::default().with_extension(".jsonc");
@@ -232,7 +232,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::ResolveOptions;
+    /// use unrspack_resolver::ResolveOptions;
     /// use std::path::{Path, PathBuf};
     ///
     /// let options = ResolveOptions::default().with_main_field("something");
@@ -249,7 +249,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::{ResolveOptions, EnforceExtension};
+    /// use unrspack_resolver::{ResolveOptions, EnforceExtension};
     /// use std::path::{Path, PathBuf};
     ///
     /// let options = ResolveOptions::default().with_force_extension(EnforceExtension::Enabled);
@@ -266,7 +266,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::{ResolveOptions};
+    /// use unrspack_resolver::{ResolveOptions};
     /// use std::path::{Path, PathBuf};
     ///
     /// let options = ResolveOptions::default().with_fully_specified(true);
@@ -283,7 +283,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::{ResolveOptions};
+    /// use unrspack_resolver::{ResolveOptions};
     /// use std::path::{Path, PathBuf};
     ///
     /// let options = ResolveOptions::default().with_prefer_relative(true);
@@ -300,7 +300,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::{ResolveOptions};
+    /// use unrspack_resolver::{ResolveOptions};
     /// use std::path::{Path, PathBuf};
     ///
     /// let options = ResolveOptions::default().with_prefer_absolute(true);
@@ -317,7 +317,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::{ResolveOptions};
+    /// use unrspack_resolver::{ResolveOptions};
     ///
     /// let options = ResolveOptions::default().with_symbolic_link(false);
     /// assert_eq!(options.symlinks, false);
@@ -333,7 +333,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::{ResolveOptions};
+    /// use unrspack_resolver::{ResolveOptions};
     ///
     /// let options = ResolveOptions::default().with_module("module");
     /// assert!(options.modules.contains(&"module".to_string()));
@@ -349,7 +349,7 @@ impl ResolveOptions {
     /// ## Examples
     ///
     /// ```
-    /// use oxc_resolver::{ResolveOptions};
+    /// use unrspack_resolver::{ResolveOptions};
     ///
     /// let options = ResolveOptions::default().with_main_file("foo");
     /// assert!(options.main_files.contains(&"foo".to_string()));
@@ -475,8 +475,7 @@ impl Default for ResolveOptions {
             main_fields: vec!["main".into()],
             main_files: vec!["index".into()],
             modules: vec!["node_modules".into()],
-            #[cfg(feature = "yarn_pnp")]
-            pnp_manifest: None,
+            enable_pnp: true,
             resolve_to_context: false,
             prefer_relative: false,
             prefer_absolute: false,
@@ -616,6 +615,8 @@ mod test {
             builtin_modules: false,
             condition_names: vec![],
             description_files: vec![],
+            #[cfg(feature = "yarn_pnp")]
+            enable_pnp: true,
             enforce_extension: EnforceExtension::Disabled,
             exports_fields: vec![],
             extension_alias: vec![],
@@ -626,8 +627,6 @@ mod test {
             main_fields: vec![],
             main_files: vec![],
             modules: vec![],
-            #[cfg(feature = "yarn_pnp")]
-            pnp_manifest: None,
             prefer_absolute: false,
             prefer_relative: false,
             resolve_to_context: false,