+++ src/tests/pnp.rs
@@ -3,6 +3,7 @@
 //! enhanced_resolve's test <https://github.com/webpack/enhanced-resolve/blob/main/test/pnp.test.js>
 //! cannot be ported over because it uses mocks on `pnpApi` provided by the runtime.
 
+use crate::ResolveError::NotFound;
 use crate::{ResolveOptions, Resolver};
 
 #[test]
@@ -11,7 +12,7 @@ fn pnp1() {
 
     let resolver = Resolver::new(ResolveOptions {
         extensions: vec![".js".into()],
-        pnp_manifest: Some(pnp::load_pnp_manifest(fixture.join(".pnp.cjs")).unwrap()),
+        condition_names: vec!["import".into()],
         ..ResolveOptions::default()
     });
 
@@ -19,6 +20,13 @@ fn pnp1() {
         resolver.resolve(&fixture, "is-even").map(|r| r.full_path()),
         Ok(fixture.join(
             ".yarn/cache/is-even-npm-1.0.0-9f726520dc-2728cc2f39.zip/node_modules/is-even/index.js"
+        ))
+    );
+
+    assert_eq!(
+        resolver.resolve(&fixture, "lodash.zip").map(|r| r.full_path()),
+        Ok(fixture.join(
+            ".yarn/cache/lodash.zip-npm-4.2.0-5299417ec8-e596da80a6.zip/node_modules/lodash.zip/index.js"
         ))
     );
 
@@ -41,5 +49,47 @@ fn pnp1() {
         Ok(fixture.join(
             ".yarn/cache/is-odd-npm-3.0.1-93c3c3f41b-89ee2e353c.zip/node_modules/is-odd/index.js"
         )),
+    );
+
+    assert_eq!(
+        resolver.resolve(&fixture, "preact").map(|r| r.full_path()),
+        Ok(fixture.join(
+            ".yarn/cache/preact-npm-10.26.4-e97fe426c5-8abf64ec6f.zip/node_modules/preact/dist/preact.mjs"
+        )),
+    );
+
+    assert_eq!(
+        resolver.resolve(&fixture, "preact/devtools").map(|r| r.full_path()),
+        Ok(fixture.join(
+            ".yarn/cache/preact-npm-10.26.4-e97fe426c5-8abf64ec6f.zip/node_modules/preact/devtools/dist/devtools.mjs"
+        )),
+    );
+}
+
+#[test]
+fn resolve_in_pnp_linked_folder() {
+    let fixture = super::fixture_root().join("pnp");
+
+    let resolver = Resolver::new(ResolveOptions {
+        extensions: vec![".js".into()],
+        condition_names: vec!["import".into()],
+        ..ResolveOptions::default()
+    });
+
+    assert_eq!(
+        resolver.resolve(&fixture, "lib/lib.js").map(|r| r.full_path()),
+        Ok(fixture.join("shared/lib.js"))
+    );
+}
+
+#[test]
+fn resolve_pnp_pkg_should_failed_while_disable_pnp_mode() {
+    let fixture = super::fixture_root().join("pnp");
+
+    let resolver = Resolver::new(ResolveOptions { enable_pnp: false, ..ResolveOptions::default() });
+
+    assert_eq!(
+        resolver.resolve(&fixture, "is-even").map(|r| r.full_path()),
+        Err(NotFound("is-even".to_string()))
     );
 }